fix: add basic auth when pulling
